+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Rename `Obsidian::MarkdownConverter` to `Obsidian::ParsedMarkdownDocument`
 - Unify `Note` and `Index` classes into `Page`. This is a breaking API change. `Parser#notes is replaced by Parse#pages`. Call `Page#is_index?`to distinguish between directory derived pages and documents.
-- Removed `Parser#table_of_contents`
+- Remove `Parser#table_of_contents` and `Parser#walk_tree`.
 
 ## [0.3.0] - 2023-07-27